Added basic plant handling.
Right click on a tile to place a pre-defined plant in it. Currently the game is set to tick every 4 seconds. If a tile has a plant in it, the tile will turn green. As the plant grows you can watch it spread to other, nearby tiles. Right now in the right situation the plant kinda spreads ridiculously, but I think that will calm down as we get some herbivores in the game.

Also, I added a specific temperature and humidity parameter to each tile. Right now the plants are designed such that if their tolerance is the same as the specific value in that tile they will do the best, if the tolerance is different than the value then they won't. Eventually having an actual tolerance might be good, but I didn't want to deal with that yet.