Huh, guess Darwin was right about that evolution thing...
Plants now have a chance to evolve to better survive in their current environment. This chance is equal to a constant times the number of plants growing in that tile that tick. Yes, I know that isn't how probability works, but I think it will allow for a higher chance of evolution than doing it "correctly" and therefore make for a more interesting game.