Mutation complete!
Plants now have a chance to randomly mutate, changing one of their parameters to something else. Also made it so that it randomly determines whether or not to check evolution or mutation first, since both actions only work if a plan has grown and remove one of the grown plants on success.