Busy, busy, busy!
Added in plant creation from the UI. Added in id generation for the animals. Renamed a parameter or two. Made it harder for plants to evolve and mutate. Changed game to tick based on number of frames as opposed to seconds. This will give more expected behavior in the result of game lag.